requirements.txt
adafruit-ampy==1.0.7
asn1crypto==0.24.0
bit==0.5.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
coincurve==12.0.0
esptool==2.7
idna==2.8
pycparser==2.19
pyserial==3.4
python-bitcoinrpc==1.0
python-dotenv==0.10.3
pyudev==0.21.0
requests==2.22.0
rshell==0.0.26
six==1.12.0
urllib3==1.25.3
flask
Flask-QRcode


# Can be more convenient for development
# -e ../../projects/bedrock
-e git+https://github.com/justinmoon/bedrock.git#egg=bedroc
